APIs/tokenmetrics.com/1.0.0/python-aiohttp-client/openapi_server/controllers/default_controller.py
from typing import List, Dict
from aiohttp import web

from openapi_server import util


async def correlation(request: web.Request, tokens=None, limit=None) -> web.Response:
    """Correlation

    Correlation

    :param tokens: 
    :type tokens: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def indices(request: web.Request, exchanges=None, time_horizon=None, low_cap=None, start_date=None, end_date=None, limit=None) -> web.Response:
    """Indices

    Indices

    :param exchanges: 
    :type exchanges: str
    :param time_horizon: 
    :type time_horizon: str
    :param low_cap: 
    :type low_cap: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def investor_grades(request: web.Request, tokens=None, start_date=None, end_date=None, limit=None) -> web.Response:
    """Investor Grades

    Investor Grades

    :param tokens: 
    :type tokens: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def market_indicator(request: web.Request, start_date=None, end_date=None, limit=None) -> web.Response:
    """Market Indicator

    Market Indicator

    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def price(request: web.Request, tokens=None, start_date=None, end_date=None, limit=None, page=None) -> web.Response:
    """Price

    Price

    :param tokens: 
    :type tokens: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str
    :param page: 
    :type page: str

    """
    return web.Response(status=200)


async def price_prediction(request: web.Request, tokens=None, _date=None, limit=None) -> web.Response:
    """Price Prediction

    Price Prediction

    :param tokens: 
    :type tokens: str
    :param _date: 
    :type _date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def quantmetrics_tier1(request: web.Request, tokens=None, _date=None, limit=None) -> web.Response:
    """Quantmetrics Tier 1

    Quantmetrics Tier 1

    :param tokens: 
    :type tokens: str
    :param _date: 
    :type _date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def quantmetrics_tier2(request: web.Request, tokens=None, _date=None, limit=None) -> web.Response:
    """Quantmetrics Tier 2

    Quantmetrics Tier 2

    :param tokens: 
    :type tokens: str
    :param _date: 
    :type _date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def resistance_support(request: web.Request, tokens=None, start_date=None, end_date=None, limit=None) -> web.Response:
    """Resistance &amp; Support

    Resistance &amp; Support

    :param tokens: 
    :type tokens: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def scenario_analysis(request: web.Request, tokens=None, limit=None) -> web.Response:
    """Scenario Analysis

    Scenario Analysis

    :param tokens: 
    :type tokens: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def sentiments(request: web.Request, tokens=None, start_date=None, end_date=None, limit=None) -> web.Response:
    """Sentiments

    Sentiments

    :param tokens: 
    :type tokens: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def tokens(request: web.Request, token_ids=None, token_names=None, token_symbols=None) -> web.Response:
    """Tokens

    Tokens

    :param token_ids: 
    :type token_ids: str
    :param token_names: 
    :type token_names: str
    :param token_symbols: 
    :type token_symbols: str

    """
    return web.Response(status=200)


async def trader_grades(request: web.Request, tokens=None, start_date=None, end_date=None, limit=None) -> web.Response:
    """Trader Grades

    Trader Grades

    :param tokens: 
    :type tokens: str
    :param start_date: 
    :type start_date: str
    :param end_date: 
    :type end_date: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)


async def trading_indicator(request: web.Request, tokens=None, limit=None) -> web.Response:
    """Trading Indicator

    Trading Indicator

    :param tokens: 
    :type tokens: str
    :param limit: 
    :type limit: str

    """
    return web.Response(status=200)
